Moving Block+Door/Button+Teleporter
introduced a moving block that hasn't got its collision method correct yet. A door/button system that keeps the door open when a player is in contact with the button.
and two teleporters for each of the player, and ends the game when both players are in the corresponding tleporter.